Update drm-msm-dsi-fix-wrong-type-in-msm_dsi_host.patch (jsc#PED-1166 jsc#PED-1168 jsc#PED-1170 jsc#PED-1218 jsc#PED-1220 jsc#PED-1222 jsc#PED-1223 jsc#PED-1225)
suse-commit: c6f0b17c418a475d52afe0efeefb5ca2d5c8ab24

+++ drivers/gpu/drm/msm/dsi/dsi_host.c
@@ -498,7 +498,7 @@ int msm_dsi_runtime_resume(struct device *dev)
 
 int dsi_link_clk_set_rate_6g(struct msm_dsi_host *msm_host)
 {
-	u32 byte_intf_rate;
+	unsigned long byte_intf_rate;
 	int ret;
 
 	DBG("Set clk rates: pclk=%d, byteclk=%lu",
